fix(api): reefactor logic
Ref. metriport/metriport-internal#365
